floatをdoubleに変更
64bitが前提のため、floatを使う意義はあまりないと考えての変更。初期化子の中で縮小変換が生じることによって、エラーも発生していた。これは初期化子ではなく、コンストラクタを呼ぶことで対処した。

また、参照の書き方の修正や、残っていたグローバル変数の削除も行った。

+++ shooter2d/bullet.cpp
@@ -11,14 +11,14 @@ static std::weak_ptr<SDL_Texture> BulletImageAdd;
 class EnemyBullet : public Bullet
 {
 public:
-	EnemyBullet(const Vector2<float>& position, std::unique_ptr<Sprite>&& sprite, std::unique_ptr<Collider>&& collider)
+	EnemyBullet(const Vector2<double>& position, std::unique_ptr<Sprite>&& sprite, std::unique_ptr<Collider>&& collider)
 		: Bullet(position, std::move(sprite), std::move(collider), EffectManager::EffectID::None, 1)
 		, sound(std::make_unique<Sound>(AssetLoader::Create().GetChunk("se/shot1.wav")))
 	{
 		sound->SetVolume(Sound::MaxVolume / 8);
 	}
 
-	void Shot(const float speed, const float angle) override
+	void Shot(const double speed, const double angle) override
 	{
 		Bullet::Shot(speed, angle);
 		sound->Played();
@@ -30,8 +30,8 @@ private:
 class LargeRedBullet : public EnemyBullet
 {
 public:
-	LargeRedBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageAdd), std::make_unique<CircleCollider>(21.0f))
+	LargeRedBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageAdd), std::make_unique<CircleCollider>(21.0))
 	{
 		clip = { 320, 0, 64, 64 };
 		sprite->SetBlendModeAdd();
@@ -41,8 +41,8 @@ public:
 class LargeBlueBullet : public EnemyBullet
 {
 public:
-	LargeBlueBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageAdd), std::make_unique<CircleCollider>(21.0f))
+	LargeBlueBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageAdd), std::make_unique<CircleCollider>(21.0))
 	{
 		clip = { 448, 0, 64, 64 };
 		sprite->SetBlendModeAdd();
@@ -52,8 +52,8 @@ public:
 class MiddleRedBullet : public EnemyBullet
 {
 public:
-	MiddleRedBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(9.0f))
+	MiddleRedBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(9.0))
 	{
 		clip = { 0, 50, 30, 30 };
 	}
@@ -62,8 +62,8 @@ public:
 class MiddleBlueBullet : public EnemyBullet
 {
 public:
-	MiddleBlueBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(9.0f))
+	MiddleBlueBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(9.0))
 	{
 		clip = { 150, 50, 30, 30 };
 	}
@@ -72,8 +72,8 @@ public:
 class SmallRedBullet : public EnemyBullet
 {
 public:
-	SmallRedBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(4.0f))
+	SmallRedBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(4.0))
 	{
 		clip = { 1, 13, 16, 16 };
 	}
@@ -82,8 +82,8 @@ public:
 class SmallBlueBullet : public EnemyBullet
 {
 public:
-	SmallBlueBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(4.0f))
+	SmallBlueBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(4.0))
 	{
 		clip = { 91, 13, 16, 16 };
 	}
@@ -92,8 +92,8 @@ public:
 class TinyRedBullet : public EnemyBullet
 {
 public:
-	TinyRedBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0f))
+	TinyRedBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0))
 	{
 		clip = { 1, 2, 9, 9 };
 	}
@@ -102,8 +102,8 @@ public:
 class TinyBlueBullet : public EnemyBullet
 {
 public:
-	TinyBlueBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0f))
+	TinyBlueBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0))
 	{
 		clip = { 61, 2, 9, 9 };
 	}
@@ -112,8 +112,8 @@ public:
 class ScaleRedBullet : public EnemyBullet
 {
 public:
-	ScaleRedBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0f))
+	ScaleRedBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0))
 	{
 		clip = { 1, 209, 15, 15 };
 	}
@@ -122,8 +122,8 @@ public:
 class ScaleBlueBullet : public EnemyBullet
 {
 public:
-	ScaleBlueBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0f))
+	ScaleBlueBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0))
 	{
 		clip = { 91, 209, 15, 15 };
 	}
@@ -132,8 +132,8 @@ public:
 class RiceRedBullet : public EnemyBullet
 {
 public:
-	RiceRedBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0f))
+	RiceRedBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0))
 	{
 		clip = { 1, 101, 9, 16 };
 	}
@@ -142,8 +142,8 @@ public:
 class RiceBlueBullet : public EnemyBullet
 {
 public:
-	RiceBlueBullet(const Vector2<float>& position)
-		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0f))
+	RiceBlueBullet(const Vector2<double>& position)
+		: EnemyBullet(position, std::make_unique<Sprite>(BulletImageBlend), std::make_unique<CircleCollider>(3.0))
 	{
 		clip = { 61, 101, 9, 16 };
 	}
@@ -152,14 +152,14 @@ public:
 class PlayerBullet : public Bullet
 {
 public:
-	PlayerBullet(const Vector2<float>& position, std::unique_ptr<Sprite>&& sprite, std::unique_ptr<Collider>&& collider)
+	PlayerBullet(const Vector2<double>& position, std::unique_ptr<Sprite>&& sprite, std::unique_ptr<Collider>&& collider)
 		: Bullet(position, std::move(sprite), std::move(collider), EffectManager::EffectID::None, 4)
 		, sound(std::make_unique<Sound>(AssetLoader::Create().GetChunk("se/sha04.wav")))
 	{
 		sound->SetVolume(Sound::MaxVolume / 32);
 	}
 
-	void Shot(const float speed, const float angle) override
+	void Shot(const double speed, const double angle) override
 	{
 		Bullet::Shot(speed, angle);
 		sound->Played();
@@ -171,8 +171,8 @@ private:
 class ReimuNormalBullet : public PlayerBullet
 {
 public:
-	ReimuNormalBullet(const Vector2<float>& position)
-		: PlayerBullet(position, std::make_unique<Sprite>(AssetLoader::Create().GetTexture("images/Shot1.png")), std::make_unique<CircleCollider>(Vector2<float>{ 0.0f, -23.0f }, 6.5f))
+	ReimuNormalBullet(const Vector2<double>& position)
+		: PlayerBullet(position, std::make_unique<Sprite>(AssetLoader::Create().GetTexture("images/Shot1.png")), std::make_unique<CircleCollider>(Vector2<double>{ 0.0, -23.0 }, 6.5))
 	{
 		clip = { 2, 3, 13, 63 };
 	}
@@ -181,8 +181,8 @@ public:
 class MarisaNormalBullet : public PlayerBullet
 {
 public:
-	MarisaNormalBullet(const Vector2<float>& position)
-		: PlayerBullet(position, std::make_unique<Sprite>(AssetLoader::Create().GetTexture("images/Shot2.png")), std::make_unique<CircleCollider>(6.5f))
+	MarisaNormalBullet(const Vector2<double>& position)
+		: PlayerBullet(position, std::make_unique<Sprite>(AssetLoader::Create().GetTexture("images/Shot2.png")), std::make_unique<CircleCollider>(6.5))
 	{
 		clip = { 1, 2, 15, 26 };
 	}
@@ -191,8 +191,8 @@ public:
 class SanaeNormalBullet : public PlayerBullet
 {
 public:
-	SanaeNormalBullet(const Vector2<float>& position)
-		: PlayerBullet(position, std::make_unique<Sprite>(AssetLoader::Create().GetTexture("images/Shot3.png")), std::make_unique<CircleCollider>(Vector2<float>{ 0.0f, -23.0f }, 6.5f))
+	SanaeNormalBullet(const Vector2<double>& position)
+		: PlayerBullet(position, std::make_unique<Sprite>(AssetLoader::Create().GetTexture("images/Shot3.png")), std::make_unique<CircleCollider>(Vector2<double>{ 0.0, -23.0 }, 6.5))
 	{
 		clip = { 0, 0, 16, 16 };
 	}
@@ -205,13 +205,13 @@ public:
 /// <param name="collider">当たり判定クラスへのポインタ</param>
 /// <param name="effectID">消滅エフェクトのID</param>
 /// <param name="damage">衝突時に相手に与えるダメージ</param>
-Bullet::Bullet(const Vector2<float>& position, std::unique_ptr<Sprite>&& sprite, std::unique_ptr<Collider>&& collider, EffectManager::EffectID effectID, unsigned int damage)
-	: Mover(position, 0.0f, M_PI_2, std::move(sprite), std::move(collider), effectID, damage, 0)
+Bullet::Bullet(const Vector2<double>& position, std::unique_ptr<Sprite>&& sprite, std::unique_ptr<Collider>&& collider, EffectManager::EffectID effectID, unsigned int damage)
+	: Mover(position, 0.0, M_PI_2, std::move(sprite), std::move(collider), effectID, damage, 0)
 {}
 
 void Bullet::Draw() const
 {
-	sprite->Draw(position, angle + M_PI_2, 1.0f);  // 元の画像は -PI/2 の向きが正位置。よって、画像の回転角は +PI/2 される。
+	sprite->Draw(position, angle + M_PI_2, 1.0);  // 元の画像は -PI/2 の向きが正位置。よって、画像の回転角は +PI/2 される。
 }
 
 void Bullet::OnCollide(Mover&)
@@ -220,7 +220,7 @@ void Bullet::OnCollide(Mover&)
 	hitPoint = 0;
 }
 
-void Bullet::Shot(const float speed, const float angle)
+void Bullet::Shot(const double speed, const double angle)
 {
 	Mover::spawned();
 	SetSpeed(speed);
@@ -244,12 +244,12 @@ BulletManager::BulletManager()
 	// 注意：IDの最初と最後に依存する。
 	for (int i = static_cast<int>(BulletID::LargeRed); i < static_cast<int>(BulletID::RiceBlue); i++) {
 		for (int j = 0; j < 50; j++) {
-			bullets.push_back(GenerateObject(static_cast<BulletID>(i), Vector2<float>{ 0.0f, 0.0f }).lock());  // 予め生成する。
+			bullets.push_back(GenerateObject(static_cast<BulletID>(i), Vector2<double>{ 0.0, 0.0 }).lock());  // 予め生成する。
 		}
 	}
 }
 
-std::weak_ptr<Bullet> BulletManager::GenerateObject(const BulletID id, const Vector2<float>& position)
+std::weak_ptr<Bullet> BulletManager::GenerateObject(const BulletID id, const Vector2<double>& position)
 {
 	std::weak_ptr<Bullet> newObject;
 	switch (id) {
