+++ shooter2d/effect.cpp
@@ -8,7 +8,7 @@ using namespace Shooter;
 class NoneEffect : public Effect
 {
 public:
-	NoneEffect(const Vector2<float>& position)
+	NoneEffect(const Vector2<double>& position)
 		: Effect(position, std::make_unique<Sprite>(AssetLoader::Create().GetTexture("images/effect_circle.png")), std::make_unique<Sound>(AssetLoader::Create().GetChunk("se/enemy_vanish_effect-A.wav")))
 	{}
 
@@ -18,7 +18,7 @@ public:
 class DefetedPlayerEffect : public Effect
 {
 public:
-	DefetedPlayerEffect(const Vector2<float>& position)
+	DefetedPlayerEffect(const Vector2<double>& position)
 		: Effect(position, std::make_unique<Sprite>(AssetLoader::Create().GetTexture("images/effect_circle.png")), std::make_unique<Sound>(AssetLoader::Create().GetChunk("se/nc899.wav")))
 	{
 		clips[0] = { 0, 128, 128, 128 };
@@ -31,7 +31,7 @@ public:
 	{
 		if (counter < AnimationFrames) {
 			alpha = static_cast<int>(255 - 255.e0 * counter / AnimationFrames);
-			scalingRate = 0.5f + counter * 0.15f;
+			scalingRate = 0.5 + counter * 0.15;
 			++counter;
 		} else {
 			enabled = false;
@@ -44,7 +44,7 @@ public:
 class CircleEffect : public Effect
 {
 public:
-	CircleEffect(const Vector2<float>& position)
+	CircleEffect(const Vector2<double>& position)
 		: Effect(position, std::make_unique<Sprite>(AssetLoader::Create().GetTexture("images/effect_circle.png")), std::make_unique<Sound>(AssetLoader::Create().GetChunk("se/enemy_vanish_effect-A.wav")))
 	{
 		this->sound->SetVolume(Sound::MaxVolume / 2);
@@ -54,7 +54,7 @@ public:
 	{
 		if (counter < AnimationFrames) {
 			alpha = static_cast<int>(255 - 255.e0 * counter / AnimationFrames);
-			scalingRate = 0.5f + counter * 0.05f;
+			scalingRate = 0.5 + counter * 0.05;
 			++counter;
 		} else {
 			enabled = false;
@@ -66,7 +66,7 @@ public:
 class RedCircleEffect : public CircleEffect
 {
 public:
-	RedCircleEffect(const Vector2<float>& position)
+	RedCircleEffect(const Vector2<double>& position)
 		: CircleEffect(position)
 	{
 		clips[0] = { 0, 128, 128, 128 };
@@ -78,7 +78,7 @@ public:
 class BlueCircleEffect : public CircleEffect
 {
 public:
-	BlueCircleEffect(const Vector2<float>& position)
+	BlueCircleEffect(const Vector2<double>& position)
 		: CircleEffect(position)
 	{
 		clips[0] = { 128, 128, 128, 128 };
@@ -89,7 +89,7 @@ public:
 
 /******************************** Effect *********************************/
 
-Effect::Effect(const Vector2<float>& position, std::unique_ptr<Sprite>&& sprite, std::unique_ptr<Sound>&& sound)
+Effect::Effect(const Vector2<double>& position, std::unique_ptr<Sprite>&& sprite, std::unique_ptr<Sound>&& sound)
 	: GameObject(false, position)
 	, sprite(std::move(sprite))
 	, sound(std::move(sound))
@@ -97,7 +97,7 @@ Effect::Effect(const Vector2<float>& position, std::unique_ptr<Sprite>&& sprite,
 
 void Effect::Draw() const
 {
-	sprite->Draw(position, 0.0f, scalingRate);
+	sprite->Draw(position, 0.0, scalingRate);
 }
 
 void Effect::Played()
@@ -105,14 +105,14 @@ void Effect::Played()
 	enabled = true;
 	counter = 0;
 	alpha = 255;
-	scalingRate = 0.0f;
+	scalingRate = 0.0;
 	sound->Played();
 }
 
 void Effect::Update()
 {
 	auto clipFromImage = [this]() -> Rect<int>& {
-		/*if (static_cast<float>(counter) / AnimationFrames < 0.3f)
+		/*if (static_cast<double>(counter) / AnimationFrames < 0.3)
 			return clips[1];
 		else
 			return clips[2];*/
@@ -125,7 +125,7 @@ void Effect::Update()
 
 /******************************** EffectManager *********************************/
 
-std::weak_ptr<Effect> EffectManager::GenerateObject(const EffectID id, const Vector2<float>& position)
+std::weak_ptr<Effect> EffectManager::GenerateObject(const EffectID id, const Vector2<double>& position)
 {
 	std::weak_ptr<Effect> newObject;
 	switch (id) {
