+++ shooter2d/media.h
@@ -12,10 +12,10 @@ namespace Shooter {
 		friend class Singleton<Media>;
 	public:
 		friend void Sprite::Draw() const;
-		friend void Sprite::Draw(const Vector2<float>&) const;
-		friend void Sprite::Draw(const Vector2<float>&, const float, const float) const;
+		friend void Sprite::Draw(const Vector2<double>&) const;
+		friend void Sprite::Draw(const Vector2<double>&, const double, const double) const;
 		friend std::weak_ptr<SDL_Texture> AssetLoader::GetTexture(const std::string);
-		friend void Label::Write(const Vector2<float>&) const;
+		friend void Label::Write(const Vector2<double>&) const;
 		friend std::weak_ptr<SDL_Texture> AssetLoader::TakeScreenshot();
 		friend class Shape;
 		friend void RectangleShape::Draw(const Vector2<int>&) const;
